+++ R/irp_predict_transmission_mir.R
@@ -122,35 +122,43 @@ irp_eac_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = mean, 
     rlang::warn(paste0("The maximum wavenumber value in `x` is ", x_flat$x[[nrow(x)]], " , but should be ", config$irp_preprocess$clip_range$end, " or larger."))
   }
 
-  # preprocess the spectra
+  # preprocessing
   x <-
-    irp_preprocess(
-      x,
-      do_interpolate = config$irp_preprocess$do_interpolate,
-      interpolate_start = config$irp_preprocess$interpolate_start,
-      interpolate_dw = config$irp_preprocess$interpolate_dw,
-      do_clip = config$irp_preprocess$do_clip,
-      clip_range = config$irp_preprocess$clip_range,
-      do_interpolate_region = config$irp_preprocess$do_interpolate_region,
-      interpolate_region_range = config$irp_preprocess$interpolate_region_range,
-      do_bc = config$irp_preprocess$do_bc,
-      bc_method = config$irp_preprocess$bc_method,
-      bc_cutoff = config$irp_preprocess$bc_cutoff,
-      bc_do_impute = config$irp_preprocess$bc_do_impute,
-      do_smooth = config$irp_preprocess$do_smooth,
-      do_normalise = config$irp_preprocess$do_normalise,
-      normalise_method = config$irp_preprocess$normalise_method,
-      do_bin = config$irp_preprocess$do_bin,
-      bin_width = config$irp_preprocess$bin_width,
-      bin_new_x_type = config$irp_preprocess$bin_new_x_type,
-      do_scale = config$irp_preprocess$do_scale,
-      scale_center = config$data_scale$x_center,
-      scale_scale = config$data_scale$x_scale,
-      do_return_as_ir = FALSE
+    irp_preprocess_eb1014(x, config = config)
+
+  # check prediction domain
+  prediction_domain <-
+    switch(
+      check_prediction_domain,
+      "train" = irpeatmodels::model_eac_1_prediction_domain$train,
+      "test" = irpeatmodels::model_eac_1_prediction_domain$train,
+      "none" = NULL
     )
 
+  res_pd <-
+    if(check_prediction_domain != "none") {
+      tibble::tibble(
+        y =
+          x %>%
+          irp_is_in_prediction_domain(prediction_domain = prediction_domain) %>%
+          dplyr::pull(.data$is_in_prediction_domain)
+      )
+    } else {
+      tibble::tibble(
+        y = rep(NA, nrow(x_or))
+      )
+    }
+
+  # reformat for predictions
+  res <- ir::ir_flatten(x)
+  res_colnames <- paste0("V", res[, 1, drop= TRUE])
+  res <- as.data.frame(t(res[, -1, drop = FALSE]))
+  attr(res, "scaled:center") <- attr(x, "scaled:center")
+  attr(res, "scaled:scale") <- attr(x, "scaled:scale")
+  colnames(res) <- res_colnames
+
   # predict
-  res <- as.data.frame(rstanarm::posterior_predict(m, newdata = data.frame(x = I(as.matrix(x)), stringsAsFactors = FALSE), ...))
+  res <- as.data.frame(rstanarm::posterior_predict(m, newdata = data.frame(x = I(as.matrix(res)), stringsAsFactors = FALSE), ...))
   res <- res * config$data_scale$y_scale + config$data_scale$y_center
 
   # summarize and add unit
@@ -163,8 +171,8 @@ irp_eac_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = mean, 
       summary_function_sd = stats::sd
     )
 
-  x_or$eac <- res
-  x_or
+  x_or$eac_1 <- res
+  cbind(x_or, res_pd %>% stats::setNames(nm = "eac_1_in_pd"))
 
 }
 
@@ -204,35 +212,43 @@ irp_edc_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = mean, 
     rlang::warn(paste0("The maximum wavenumber value in `x` is ", x_flat$x[[nrow(x)]], " , but should be ", config$irp_preprocess$clip_range$end, " or larger."))
   }
 
-  # preprocess the spectra
+  # preprocessing
   x <-
-    irp_preprocess(
-      x,
-      do_interpolate = config$irp_preprocess$do_interpolate,
-      interpolate_start = config$irp_preprocess$interpolate_start,
-      interpolate_dw = config$irp_preprocess$interpolate_dw,
-      do_clip = config$irp_preprocess$do_clip,
-      clip_range = config$irp_preprocess$clip_range,
-      do_interpolate_region = config$irp_preprocess$do_interpolate_region,
-      interpolate_region_range = config$irp_preprocess$interpolate_region_range,
-      do_bc = config$irp_preprocess$do_bc,
-      bc_method = config$irp_preprocess$bc_method,
-      bc_cutoff = config$irp_preprocess$bc_cutoff,
-      bc_do_impute = config$irp_preprocess$bc_do_impute,
-      do_smooth = config$irp_preprocess$do_smooth,
-      do_normalise = config$irp_preprocess$do_normalise,
-      normalise_method = config$irp_preprocess$normalise_method,
-      do_bin = config$irp_preprocess$do_bin,
-      bin_width = config$irp_preprocess$bin_width,
-      bin_new_x_type = config$irp_preprocess$bin_new_x_type,
-      do_scale = config$irp_preprocess$do_scale,
-      scale_center = config$data_scale$x_center,
-      scale_scale = config$data_scale$x_scale,
-      do_return_as_ir = FALSE
+    irp_preprocess_eb1014(x, config = config)
+
+  # check prediction domain
+  prediction_domain <-
+    switch(
+      check_prediction_domain,
+      "train" = irpeatmodels::model_edc_1_prediction_domain$train,
+      "test" = irpeatmodels::model_edc_1_prediction_domain$train,
+      "none" = NULL
     )
 
+  res_pd <-
+    if(check_prediction_domain != "none") {
+      tibble::tibble(
+        y =
+          x %>%
+          irp_is_in_prediction_domain(prediction_domain = prediction_domain) %>%
+          dplyr::pull(.data$is_in_prediction_domain)
+      )
+    } else {
+      tibble::tibble(
+        y = rep(NA, nrow(x_or))
+      )
+    }
+
+  # reformat for predictions
+  res <- ir::ir_flatten(x)
+  res_colnames <- paste0("V", res[, 1, drop= TRUE])
+  res <- as.data.frame(t(res[, -1, drop = FALSE]))
+  attr(res, "scaled:center") <- attr(x, "scaled:center")
+  attr(res, "scaled:scale") <- attr(x, "scaled:scale")
+  colnames(res) <- res_colnames
+
   # predict
-  res <- as.data.frame(rstanarm::posterior_predict(m, newdata = data.frame(x = I(as.matrix(x)), stringsAsFactors = FALSE), ...))
+  res <- as.data.frame(rstanarm::posterior_predict(m, newdata = data.frame(x = I(as.matrix(res)), stringsAsFactors = FALSE), ...))
   res <- res * config$data_scale$y_scale + config$data_scale$y_center
 
   # summarize and add unit
@@ -245,8 +261,8 @@ irp_edc_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = mean, 
       summary_function_sd = stats::sd
     )
 
-  x_or$edc <- res
-  x_or
+  x_or$edc_1 <- res
+  cbind(x_or, res_pd %>% stats::setNames(nm = "eac_1_in_pd"))
 
 }
 
@@ -858,8 +874,8 @@ irp_porosity_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = m
     irp_bulk_density_1(..., do_summary = FALSE, check_prediction_domain = check_prediction_domain) %>%
     dplyr::rename(non_macroporosity_1_in_pd = "bulk_density_1_in_pd") %>%
     dplyr::mutate(
-      macroporosity_1_in_pd = non_macroporosity_1_in_pd,
-      volume_fraction_solids_1_in_pd = non_macroporosity_1_in_pd
+      macroporosity_1_in_pd = .data$non_macroporosity_1_in_pd,
+      volume_fraction_solids_1_in_pd = .data$non_macroporosity_1_in_pd
     )
 
   ## predict porosity
@@ -900,14 +916,14 @@ irp_porosity_1 <- function(x, ..., do_summary = FALSE, summary_function_mean = m
     purrr::map(seq_len(dim(mu_inv)[[2]]), function(i) {
       brms::rdirichlet(n = dim(mu_inv)[[1]], alpha = mu_inv[, i, ] * m_draws_porosity_1$phi) %>%
         as.data.frame() %>%
-        setNames(nm = c("volume_fraction_solids_1", "non_macroporosity_1", "macroporosity_1"))
+        stats::setNames(nm = c("volume_fraction_solids_1", "non_macroporosity_1", "macroporosity_1"))
     }) %>%
     purrr::transpose() %>%
     purrr::map2_dfc(names(.), function(.x, .y) {
       tibble::tibble(
         y = unname(irp_summarize_predictions(as.data.frame(.x), x_unit = "L/L", do_summary = do_summary, summary_function_mean = summary_function_mean, summary_function_sd = summary_function_sd))
       ) %>%
-        setNames(nm = .y)
+        stats::setNames(nm = .y)
     })
 
   cbind(x_or, res, x %>% dplyr::select(.data$non_macroporosity_1_in_pd, .data$macroporosity_1_in_pd, .data$volume_fraction_solids_1_in_pd))
@@ -1081,7 +1097,7 @@ irp_saturated_hydraulic_conductivity_1 <- function(x, ..., do_summary = FALSE, s
 
   res <-
     tibble::tibble(y = res) %>%
-    setNames(nm = "saturated_hydraulic_conductivity_1")
+    stats::setNames(nm = "saturated_hydraulic_conductivity_1")
 
   cbind(x_or, res, x %>% dplyr::select(.data$saturated_hydraulic_conductivity_1_in_pd))